Cock and ball torture
From Wikipedia, the free encyclopedia

at en.wikipedia.org/

Cock and ball torture (CBT) is a sexual activity involving torture of the male genitals. This may involve directly painful activities, such as wax play, genital spanking, squeezing, ball-busting, genital flogging, urethral play, tickle torture, erotic electrostimulation or even kicking. The recipient of such activities may receive direct physical pleasure via masochism, from knowledge that the play is pleasing to a sadistic dominant.

Image: Electrostimulation applied on a penis.

Contents

Section 1: In pornography

Section 2: Ball stretcher

Section 3: Parachute

Section 4: Humbler

Section 5: Testicle Cuff

Section 1: In pornography

In occasion to its occasional role in BDSM pornography, tamakeri (lit. ball kicking) is a separate genre in Japan. One notable actress in tamakeri is Erika Nagai, who typically uses her marital art skills to knee or kick man in the testicles.

Section 2: Ball stretcher

A ball stretcher is a sex toy that is fastened around a man in order to elongate the scrotum and provide a feel of weight pulling the testicles away from the body. While leather stretcher are most common, other models are made of steel rings that are fastened with screws, causing additional mildly uncomfortable weight to the wearer. The length of the stretcher may vary from 1-4 inches, and the steel models can weight as much as five pounds.

Section 3: Parachute

A parachute is a small collar, usually made from leather, which fastens around the scrotum, and from which weights can be hung. Conical in shape, with three or four short chains hanging beneath, to which weights can be applied. Used as part of cock and ball torture within a BDSM relationship, the parachute provides a constant drag, and squeezing effet on the man's testicles, moderate weights from 3-5 kg can be suspended, especially during bondage. Smaller weights can be used when the man is free to move, when the swinging effect of the weight can restrict some movements, as well as providing a visual stimulus for the dominant partner.

Section 4: Humbler

A humbler is a BDSM physical restraint device used to restraint the movement of a submissive male partner in a BDSM scene. The humbler consists of a testicle cuff device that clamps around the base of the scrotum, mounted in the center of a bar that passes behind the thighs at the base of the buttocks. This forces the wearer to keep his legs forward, as any attempt to straighten the legs even slightly pulls directly on the scrotum, causing from considerable discomfort to extreme pain.

Section 5: Testicle cuff

A testicle cuff is a ring shaped device around the scrotum between the body and the testic-