Horse cock is fucking mediocre
Okay, you know what? Im gonna say it.

Horse cock is fucking mediocre. You know why?

Because it lacks complexity. Canine cock has that glorious knot

"But horsecock has a flare!!!!" Yeah and after it cums its soft. Who wants a
partner with super short stamina? "But the flare keeps the cum inside you!!!"
For what? 15 seconds!? A knot does that too, for much longer, and its in just
the right place to press against your prostate/g-spot. So with each one of your
lovers heartbeats you can get propelled further into pleasure.

Or what about the luscious curved taper of a dolphin or orca? Thick at the base
and the taper makes it an easy adjustment for beginners wroldwide. Not to
mention they're PREHENSILE.

And hows about the barbed feline cock? It takes some getting used to for sure,
and its not for everyone. But those barbs can really take you to the next
level, the pain mixed in with the pleasure is exquisite.

Look at snakes and sharks! There's two dicks to play with.

Dragons have bumps and ridges, tapered and knotted, and every combination
between!!!

Alligators and crocodiles have dicks that are literally ALWAYS erect.

And don't even start on size. Last I checked, sex isn't all about the dick, as
a matter of fact you could say the cock is less than half of the important part
of sex.

Horsecock is just glorified, uncomfortable, human cock, with no merits other
than size, and thats a pretty fucking weak merit for this lauded "god of all
cocks".